+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 - Add conversion from `publicKeyJwk` feature via `from()`.
 
+### Fixed
+- Allow `@context` array values in multikeys.
+
 ## 1.1.1 - 2024-01-11
 
 ### Fixed